requirements.txt
https://github.com/django/django/archive/1.6b1.zip#egg=Django==1.6b1
PyExecJS==1.0.4
South==0.8.1
amqp==1.0.13
anyjson==0.3.3
billiard==2.7.3.31
boto==2.9.9
celery==3.0.21
coffee-compressor-compiler==0.2.0
distribute==0.6.34
dj-database-url==0.2.2
django-appconf==0.6
django-celery==3.0.21
django-celery-with-redis==3.0
django-compressor==1.3
django-debug-toolbar==0.9.4
django-s3-folder-storage==0.1
django-storages==1.1.8
django-tables2==0.13.0
ember-compressor-compiler==0.3.1
gunicorn==17.5
kombu==2.5.12
psycopg2==2.5.1
python-dateutil==2.1
pytz==2013b
redis==2.7.6
six==1.3.0
sorl-thumbnail==11.12
wsgiref==0.1.2
ipython
